Started testing Game WIP
